+++ sqlparser/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## unreleased
+- Support common table expressions
 - Handle special `rowid`, `oid`, `__rowid__` references
 - Support references to `sqlite_master` and `sqlite_sequence` tables